fix: redirect 유효성 검사 오류 수정
