+++ src/components/routes/AuthRoute.tsx
@@ -1,6 +1,6 @@
-import { matchPath, Navigate, Outlet, useLocation } from "react-router-dom";
+import { Navigate, Outlet, useLocation } from "react-router-dom";
 import { getCookieValue } from "@/utils/cookie";
-import { ROUTE_PATH } from "@/components/routes/Routes";
+import { checkValidPath, ROUTE_PATH } from "./routePath";
 import { AUTH_COOKIE_KEY, useAuth } from "@/contexts/authContext";
 
 type AuthRouteProps = {
@@ -29,10 +29,3 @@ const AuthRoute = ({ required = false }: AuthRouteProps) => {
 };
 
 export default AuthRoute;
-
-const checkValidPath = (path: string): boolean => {
-  return Object.values(ROUTE_PATH).some((pattern) => {
-    const match = matchPath({ path: pattern, end: true }, path);
-    return match !== null;
-  });
-};
